+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- More shorthand BotBuilder extensions such as `onChannelCreate`, `onMessageEdit`, etc
+### Changed
+- Rename some events for consistency
+- Rename some BotBuilder extensions for consistency
+
 ## [0.2.1] - 2019-09-03
 ### Changed
 - Updated stack to Kotlin 1.3.50